requirements.txt
rackspace-monitoring==0.6.2
dnspython3==1.11.1
python-nmap==0.3.4
